pyproject.toml
[project]
name = "open-claude"
version = "0.10.0"
description = "Unofficial open source toolkit for Claude Code — AI-powered business operating system"
requires-python = ">=3.10"
dependencies = [
    "rich>=13.0",
    "schedule>=1.2",
    "flask>=3.0",
    "flask-cors>=4.0",
    "flask-sqlalchemy>=3.1",
    "flask-login>=0.6",
    "python-dotenv>=1.0",
    "bcrypt>=4.0",
    "requests>=2.31",
    "pyyaml>=6.0",
    "flask-sock>=0.7",
]

[tool.uv]
dev-dependencies = []
